+++ src/store/indices.ts
@@ -2,6 +2,7 @@ import { defineStore } from 'pinia'
 import { DEFAULT_HIDE_INDICES_REGEX } from '../consts'
 
 type IndicesState = {
+  filter: string,
   showHiddenIndices: boolean,
   stickyTableHeader: boolean,
   hideIndicesRegex: string
@@ -9,6 +10,7 @@ type IndicesState = {
 
 export const useIndicesStore = defineStore('indices', {
   state: (): IndicesState => ({
+    filter: '',
     showHiddenIndices: false,
     stickyTableHeader: false,
     hideIndicesRegex: DEFAULT_HIDE_INDICES_REGEX