+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 * fixes issue with old versions of elasticsearch that do not provide a uuid, fixes [#182][i182]
 * adds document meta information when editing documents, fixes [#187][i187]
+* remap hotkey for sending rest request to `Ctrl+Enter`, fixes [#183][i183]
 
 [i182]: https://github.com/cars10/elasticvue/issues/182
 [i187]: https://github.com/cars10/elasticvue/issues/187
+[i183]: https://github.com/cars10/elasticvue/issues/183
 
 ## 1.0.1